Move all dependency related classes into inspec/dependencies
