+++ lib/inspec/dependencies/requirement.rb
