+++ lib/inspec/dependencies/vendor_index.rb
