+++ test/unit/mock/profiles/dependencies/inheritance/controls/example.rb
