+++ test/unit/mock/profiles/dependencies/profile_a/controls/example.rb
