+++ test/unit/mock/profiles/dependencies/profile_b/controls/example.rb
